Amélioration de l'affichage côté backend

+++ Back_office/submit_suppression_offre.php
@@ -8,9 +8,12 @@ exit;
 require_once(__DIR__ . '/../Frontoffice/fonctions.php');
 require_once(__DIR__ . '/../Frontoffice/connexionBDD.php');
 
-$deleteoffre= $mysqlClient->prepare('DELETE FROM offres WHERE Id = :Id');
+if (isset($_POST['id_offre']) && !empty($_POST['id_offre'])) {
+
+    $deleteoffre= $mysqlClient->prepare('DELETE FROM offres WHERE Id = :Id');
     $deleteoffre->execute([
         'Id' => $_POST['id_offre'],
-    ]);
+        ]);
     redirectToUrl('offre.php'); 
+}
 ?>
